sagemaker/requirements.txt
attrs==21.4.0
boto3==1.24.35
botocore==1.27.35
dill==0.3.5.1
google-pasta==0.2.0
importlib-metadata==4.12.0
jmespath==1.0.1
joblib==1.1.0
multiprocess==0.70.13
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pathos==0.2.9
pox==0.3.1
ppft==1.7.6.5
protobuf==3.20.1
protobuf3-to-dict==0.1.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
s3transfer==0.6.0
sagemaker==2.100.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
smdebug-rulesconfig==1.0.1
threadpoolctl==3.1.0
urllib3==1.26.10
zipp==3.8.1
tensorflow
keras
nltk